Put Tests in Classes and Added Exit Button
Put Tests in Classes and Added Exit Button
