--- UnitTestTeamDev/UnitTest2.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace UnitTestTeamDev
-{
-    [TestClass]
-    public class UnitTest2
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
