+++ UnitTestTeamDev/UnitTestTeamDev.csproj
@@ -68,17 +68,10 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="InventoryCreate.cs" />
+    <Compile Include="InventoryUnitTests.cs" />
+    <Compile Include="OrdersUnitTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CustomerUnitTests.cs" />
-    <Compile Include="OrderSelect.cs" />
-    <Compile Include="OrderCreate.cs" />
-    <Compile Include="OrderUpdate.cs" />
-    <Compile Include="OrderDelete.cs" />
-    <Compile Include="InventoryDelete.cs" />
-    <Compile Include="InventorySelect.cs" />
-    <Compile Include="InventoryUpdate.cs" />
-    <Compile Include="UnitTest2.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
